+++ app/javascript/store/modules/songs.js
@@ -1,4 +1,4 @@
-import axios from '../axios'
+import axios from '../../api/axios'
 
 const state = {
   all: [],
